Version 3.3a - Texture Streaming / Bug Fixes
This is probably one of the last bigger changes, as we're beginning to approach the limit of gains before we start to encounter bugs or unplayable graphics. Obviously, we'll learn more about the engine over time and gradually move closer, but this will be be one of the last bigger, frequent updates.

Main:
-Disable autoexposure speed rather  than increasing it. This disables the lighting change that happens when going inside buildings. Even at a soft value of 2 (default 1), it looked janky.
-Texture streaming changes that should benefit most users is enabled here
-Extract telemetry settings into their own category

Toaster:
-Further drop static shadow res
-Texture streaming changes for potato PCs