Version 5.1.1: Less ugly
Sets stream_memory to a conservative 32 MB to render detail. This shouldn't cause a hit to NVENC users, users with gpus with <1 GB RAM that stream with NVENC might want to leave this value at 0, but I think there would be bigger issues with streaming on that than the 32 MB budget.